requirements.txt
requests
psycopg2-binary
scispacy
spacy
sklearn
streamlit
neo4j
pandas
biopython
